requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.2.post1
argcomplete==3.5.2
arrow==1.3.0
bcrypt==3.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==44.0.0
dnspython==2.7.0
docker==7.1.0
dockerpty==0.4.1
dokcer==0.8
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.5
fastapi-cli==0.0.5
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
humanize==4.11.0
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
orjson==3.10.11
packaging==24.2
passlib==1.7.4
pluggy==1.5.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.9.2
pydantic-extra-types==2.10.0
pydantic-settings==2.6.1
pydantic_core==2.23.4
Pygments==2.18.0
pystache==0.6.6
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.17
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
rsa==4.9
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typer==0.13.1
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.3.0
uvicorn==0.32.0
uvloop==0.21.0
watchfiles==0.24.0
websockets==14.1